Support for specifying explicit pageant keyfile.
Instead of walking through all identities pageant can offer, we only pick the identity which matches the public key specified by an optional path to the putty keyfile. We fall back to the previous behavior if there is no file give.
Using the identity filtering we no longer risk having too many authentication failures (and thus a disconnect by the server) if our identity is not on top of the list returned by pageant.